backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
anthropic==0.26.0
weasyprint==62.3
pydantic==2.7.1
python-multipart==0.0.9
Jinja2==3.1.4
